Splitting files for better management.
