+++ io_cryengine_importer/.vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "python.pythonPath": "C:\\Users\\Geoff\\Anaconda3\\python.exe"
+}
